% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_setup.R
\name{create_project_folder}
\alias{create_project_folder}
\title{Create project directory and folder system.}
\usage{
create_project_folder(project_name, base_path = ".")
}
\description{
\code{create_project_folder()} creates a .Rproj object in a new folder at the specified directory, 
and then sets up a default directory structure.
}
