% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_setup.R
\name{grab_homeworks}
\alias{grab_homeworks}
\title{Get homeworks.}
\usage{
grab_homeworks(variables)
}
\description{
\code{grab_homeworks()} downloads a set of homework RMarkdown files from a specified URL.
}
